+++ src/packages/console/src/components/confirmationDialog.tsx
@@ -3,16 +3,18 @@ import React from 'react';
 import { Button, Direction, PaddingSize, Stack, Text, TextAlignment } from '@kibalabs/ui-react';
 import { Dialog } from './dialog';
 
-export interface IMessageDialogProps {
+export interface IConfirmationDialogProps {
   isOpen: boolean;
   title: string;
   message: string;
   confirmButtonText?: string;
+  cancelButtonText?: string;
   onConfirmClicked: () => void;
   onCloseClicked: () => void;
 }
 
-export const MessageDialog = (props: IMessageDialogProps): React.ReactElement => {
+export const ConfirmationDialog = (props: IConfirmationDialogProps): React.ReactElement => {
+  const cancelButtonText = props.cancelButtonText || 'Cancel';
   const confirmButtonText = props.confirmButtonText || 'Confirm';
 
   return (
@@ -23,7 +25,7 @@ export const MessageDialog = (props: IMessageDialogProps): React.ReactElement =>
         <Stack.Item gutterBefore={PaddingSize.Default}>
           <Stack direction={Direction.Horizontal} shouldAddGutters={true} defaultGutter={PaddingSize.Wide}>
             <Stack.Item growthFactor={1} shrinkFactor={1} />
-            <Button variant='secondary' onClicked={props.onCloseClicked} text='Cancel' />
+            <Button variant='secondary' onClicked={props.onCloseClicked} text={cancelButtonText} />
             <Button variant='primary' onClicked={props.onConfirmClicked} text={confirmButtonText} />
             <Stack.Item growthFactor={1} shrinkFactor={1} />
           </Stack>